+++ Piktosaur/ViewModels/BaseViewModel.cs
@@ -8,23 +8,25 @@ using System.Threading.Tasks;
 
 namespace Piktosaur.ViewModels
 {
+    /// <summary>
+    /// Basic implementation to add {x:Bind} support in XAML files.
+    /// This only implements properties, and not commands: this project
+    /// calls commands from the view classes directly.
+    /// </summary>
     public class BaseViewModel : INotifyPropertyChanged
     {
         public event PropertyChangedEventHandler? PropertyChanged;
 
         protected virtual void OnPropertyChanged([CallerMemberName] string propertyName = "")
         {
-            if (propertyName == "")
-            { return; }
+            if (propertyName == "") return;
             PropertyChanged?.Invoke(this, new PropertyChangedEventArgs(propertyName));
         }
 
         protected bool SetProperty<T>(ref T field, T value, [CallerMemberName] string propertyName = "")
         {
-            if (propertyName == "")
-                return false;
-            if (EqualityComparer<T>.Default.Equals(field, value))
-                return false;
+            if (propertyName == "") return false;
+            if (EqualityComparer<T>.Default.Equals(field, value)) return false;
 
             field = value;
             OnPropertyChanged(propertyName);
